+++ com.delivery-signal.eureka.client.delivery/src/main/java/com/delivery_signal/eureka/client/delivery/presentation/controller/DeliveryController.java
@@ -6,7 +6,7 @@ import com.delivery_signal.eureka.client.delivery.application.dto.DeliveryQueryR
 import com.delivery_signal.eureka.client.delivery.application.service.DeliveryService;
 import com.delivery_signal.eureka.client.delivery.presentation.dto.ApiResponse;
 import com.delivery_signal.eureka.client.delivery.presentation.dto.request.DeliveryCreateRequest;
-import com.delivery_signal.eureka.client.delivery.presentation.dto.response.PagedDeliveryResponse;
+import com.delivery_signal.eureka.client.delivery.application.dto.PagedDeliveryResponse;
 import com.delivery_signal.eureka.client.delivery.presentation.mapper.DeliveryPresentationMapper;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Sort;
